requirements.txt
fastapi
uvicorn
python-dotenv
alembic
pymysql